# Conflicts:
#	ListOfHotels.xcodeproj/project.pbxproj
#	ListOfHotels/Hotel.swift
#	ListOfHotels/HotelParser.swift
#	ListOfHotelsTests/HotelTests.swift
#	Podfile
#	Podfile.lock
#	Pods/Manifest.lock
#	Pods/Pods.xcodeproj/project.pbxproj
#	Pods/Target Support Files/ComponentKit/ComponentKit.xcconfig
#	Pods/Target Support Files/Pods/Pods-frameworks.sh
#	Pods/Target Support Files/Pods/Pods.debug.xcconfig
#	Pods/Target Support Files/Pods/Pods.release.xcconfig
#	Pods/Target Support Files/SDWebImage/SDWebImage.xcconfig

+++ ListOfHotels/Hotel.swift
@@ -21,7 +21,7 @@ public class Hotel : NSObject, Mappable {
     var promoPrice: String?
     var locality: String?
     var lon: Float?
-    var starRating: Int?
+    var starRating: Float?
     var avgPriceDescription: String?
     var location: String?
     var postalCode: String?
